Initial cleanup after refactor, moved ImageMetaInfo one level up
