+++ src/Core/Entity/Image/InputImage.php
@@ -4,7 +4,7 @@ namespace Core\Entity\Image;
 
 use Core\Entity\OptionsBag;
 use Core\Exception\ReadFileException;
-use Core\Entity\Image\ImageMetaInfo;
+use Core\Entity\ImageMetaInfo;
 
 class InputImage
 {
