+++ src/Core/Processor/ImageProcessor.php
@@ -3,6 +3,7 @@
 namespace Core\Processor;
 
 use Core\Entity\Image\OutputImage;
+use Core\Entity\ImageMetaInfo;
 
 /**
  * Class ImageProcessor
@@ -20,16 +21,10 @@ use Core\Entity\Image\OutputImage;
 class ImageProcessor extends Processor
 {
     /**
-     * This holds the width and height dimensions in pixels of the source image
-     * @var array
-     */
-    protected $sourceDimensions = [];
-
-    /**
      * Basic source image info parsed from IM identify command
-     * @var array
+     * @var ImageMetaInfo
      */
-    protected $sourceInfo = [];
+    protected $sourceImageInfo;
 
     /**
      * OptionsBag from the request
@@ -37,6 +32,10 @@ class ImageProcessor extends Processor
      */
     protected $options;
 
+    /**
+     * stores/caches image related data like dimensions
+     * @var array
+     */
     protected $geometry;
 
     /**
@@ -49,7 +48,7 @@ class ImageProcessor extends Processor
      */
     public function processNewImage(OutputImage $outputImage): OutputImage
     {
-        $this->outputImage = $outputImage;
+        $this->sourceImageInfo = $outputImage->getInputImage()->getSourceImageInfo();
         $this->options = $outputImage->getInputImage()->getOptionsBag();
         $this->generateCmdString($outputImage);
         $this->execute($outputImage->getCommandString());
@@ -65,7 +64,6 @@ class ImageProcessor extends Processor
     public function generateCmdString(OutputImage $outputImage)
     {
         // we will categorize the operation in this method and call the adecuate functions depending on the parameters
-        // 
         // first we get the data we need
         $width = $this->options->getOption('width');
         $height = $this->options->getOption('height');
@@ -74,49 +72,28 @@ class ImageProcessor extends Processor
         $command = [];
         $command[] = self::IM_CONVERT_COMMAND;
 
-        // if width AND height AND crop are defined we need to do further checks to define the type of operation we will do
+        // if width AND height AND crop are defined we need check further to define the type of operation we will do
         if ($width && $height &&  $crop) {
             $size = $this->generateCropSize();
-        } else if($width || $height) {
+        } elseif ($width || $height) {
             $size = $this->generateSimpleSize();
         }
         
-        
-        $strip = $outputImage->extract('strip');
-        $thread = $outputImage->extract('thread');
-        $resize = $outputImage->extract('resize');
-        $frame = $outputImage->extract('gif-frame');
-
-        //list($size, $extent, $gravity) = $this->generateSize($outputImage);
-
-        // we default to thumbnail
-        //$resizeOperator = $resize ? 'resize' : 'thumbnail';
-        
-        $tmpFileName = $outputImage->getInputImage()->getSourceImagePath();
-
-        //Check the source image is gif
         if ($outputImage->isInputGif()) {
             $command[] = '-coalesce';
-            if ($outputImage->getOutputImageExtension() != OutputImage::EXT_GIF) {
-                $tmpFileName .= '['.escapeshellarg($frame).']';
-            }
         }
 
-        $command[] = " " . $tmpFileName;
+        $command[] = $this->getSourceImagePath($outputImage);
         $command[] = $size;
         $command[] = ' -colorspace sRGB';
 
-        foreach ($outputImage->getInputImage()->getOptionsBag() as $key => $value) {
-            if (!empty($value) && !in_array($key, self::EXCLUDED_IM_OPTIONS)) {
-                //$command[] = "-{$key} ".escapeshellarg($value);
-            }
-        }
-
         // strip is added internally by ImageMagick when using -thumbnail
+        $strip = $outputImage->extract('strip');
         if (!empty($strip)) {
-            $command[] = "-strip ";
+            $command[] = "-strip";
         }
 
+        $thread = $outputImage->extract('thread');
         if (!empty($thread)) {
             $command[] = "-limit thread ".escapeshellarg($thread);
         }
@@ -125,6 +102,58 @@ class ImageProcessor extends Processor
 
         $commandStr = implode(' ', $command);
         $outputImage->setCommandString($commandStr);
+    }
+
+    /**
+     * IF we crop we need to know if the source image is bigger or smaller than the target size.
+     * @return string command section for the resizing.
+     *
+     * note: The shorthand version of resize to fill space will always fill the space even if image is bigger
+     */
+    protected function generateCropSize(): string
+    {
+        $this->updateTargetDimensions();
+        $command = [];
+        $command[] = $this->getResizeOperator();
+        $command[] = $this->getDimensions() . '^';
+        $command[] = '-gravity ' . $this->options->getOption('gravity');
+        $command[] = '-extent ' . $this->getDimensions();
+        return implode(' ', $command);
+    }
+
+    /**
+     * IF we simply resize we let IM deal with the calculations
+     * @return string command section for the resizing.
+     */
+    protected function generateSimpleSize(): string
+    {
+        $command = [];
+        $command[] = $this->getResizeOperator();
+        $command[] = $this->getDimensions() .
+            ($this->options->getOption('preserve-natural-size') ? escapeshellarg('>') : '');
+        return implode(' ', $command);
+    }
+
+    /**
+     * Gets the source image path and adds any extra modifiers to the string
+     * @param  OutputImage $outputImage
+     * @return string                   Path of the source file to be used in the conversion command
+     */
+    protected function getSourceImagePath(OutputImage $outputImage): string
+    {
+        $tmpFileName = $this->sourceImageInfo->getPath();
+
+        //Check the source image is gif
+        if ($outputImage->isInputGif()) {
+            $frame = $this->options->getOption('gif-frame');
+
+            // set the frame if the output image is not gif (to get ony one  frame)
+            if ($outputImage->getOutputImageExtension() !== OutputImage::EXT_GIF) {
+                $tmpFileName .= '['.escapeshellarg($frame).']';
+            }
+        }
+
+        return $tmpFileName;
     }
 
     /**
@@ -161,41 +190,6 @@ class ImageProcessor extends Processor
     }
 
     /**
-     * IF we crop we need to know if the source image is bigger or smaller than the target size.
-     * @return string command section for the resizing.
-     *
-     * 👉 docker exec -it flyimg /usr/bin/convert web/landscape-color-squares-300x200.png -resize 400x50^ -gravity center -extent 400x50 100_t15.png
-     * identify: ./100_t15.png PNG 400x50 400x50+0+0 8-bit sRGB 6.43KB 0.000u 0:00.000
-     * 
-     * 👉 docker exec -it flyimg /usr/bin/convert web/landscape-color-squares-300x200.png -resize 400x50^\> -gravity center -extent 400x50 100_t16.png
-     *  identify: ./100_t16.png PNG 400x50 400x50+0+0 8-bit sRGB 5.63KB 0.010u 0:00.009
-     *
-     *  The shorthand version of resize to fill space will always fill the space even if image is bigger
-     */
-    protected function generateCropSize(): string
-    {
-        $this->updateTargetDimensions();
-        $command = [];
-        $command[] = $this->getResizeOperator();
-        $command[] = $this->getDimensions() . '^';
-        $command[] = '-gravity ' . $this->options->getOption('gravity');
-        $command[] = '-extent ' . $this->getDimensions();
-        return implode(' ', $command);
-    }
-
-    /**
-     * IF we simply resize we let IM deal with the calculations
-     * @return string command section for the resizing.
-     */
-    protected function generateSimpleSize(): string
-    {
-        $command = [];
-        $command[] = $this->getResizeOperator();
-        $command[] = $this->getDimensions() . ($this->options->getOption('preserve-natural-size') ? escapeshellarg('>') : '');
-        return implode(' ', $command);
-    }
-
-    /**
      * This works as a cache for calculations
      * @param  string $key       the key with wich we store a calculated value
      * @param  func   $calculate function that returns a calculated value
@@ -203,7 +197,7 @@ class ImageProcessor extends Processor
      */
     protected function getGeometry($key, $calculate):string
     {
-        if(isset($this->geometry[$key])) {
+        if (isset($this->geometry[$key])) {
             return $this->geometry[$key];
         }
         $this->geometry[$key] = call_user_func($calculate);
@@ -243,8 +237,8 @@ class ImageProcessor extends Processor
 
         $targetWidth = $this->options->getOption('width');
         $targetHeight = $this->options->getOption('height');
-        $originalWidth = $this->getSourceImageDimensions()['width'];
-        $originalHeight = $this->getSourceImageDimensions()['height'];
+        $originalWidth = $this->sourceImageInfo->getDimensions()['width'];
+        $originalHeight = $this->sourceImageInfo->getDimensions()['height'];
         
         if ($originalWidth < $targetWidth) {
             $this->options->setOption('width', $originalWidth);
@@ -305,31 +299,5 @@ class ImageProcessor extends Processor
         }
 
         return [$size, $extent, $gravity];
-    }
-
-    /**
-     * @return array
-     */
-    protected function getSourceImageDimensions()
-    {
-        if (!empty($this->sourceDimensions)) {
-            return $this->sourceDimensions;
-        }
-
-        $this->sourceDimensions = $this->outputImage->getInputImage()->getSourceImageInfo()->getDimensions();
-        return $this->sourceDimensions;
-    }
-
-    /**
-     * @return array Associative array with basic image
-     */
-    protected function getSourceImageInfo()
-    {
-        if (!empty($this->sourceInfo)) {
-            return $this->sourceInfo;
-        }
-
-        $this->sourceInfo = $this->outputImage->getInputImage()->getSourceImageInfo()->getInfo();
-        return $this->sourceInfo;
     }
 }
