+++ tests/Core/Entity/ImageMetaInfoTest.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace Tests\Core\Entity\Image;
+namespace Tests\Core\Entity;
 
-use Core\Entity\Image\ImageMetaInfo;
+use Core\Entity\ImageMetaInfo;
 use Core\Entity\Image\OutputImage;
 use Core\Exception\ExecFailedException;
 use Tests\Core\BaseTest;
