+++ tests/Core/Processor/ImageProcessorTest.php
@@ -4,7 +4,7 @@ namespace Tests\Core\Processor;
 
 use Core\Processor\ImageProcessor;
 use Core\Entity\Image\OutputImage;
-use Core\Entity\Image\ImageMetaInfo;
+use Core\Entity\ImageMetaInfo;
 use Tests\Core\BaseTest;
 
 /**
@@ -66,48 +66,56 @@ class ImageProcessorTest extends BaseTest
             ['w_300', '300x200', self::PNG_TEST_LANDSCAPE_IMAGE],
             'Resize to width portrait' =>
             ['w_300', '300x451', self::PNG_TEST_PORTRAIT_IMAGE],
+
             'Resize to height square' =>
             ['h_300', '300x300', self::PNG_TEST_SQUARE_IMAGE],
             'Resize to height landscape' =>
             ['h_300', '450x300', self::PNG_TEST_LANDSCAPE_IMAGE],
             'Resize to height portrait' =>
             ['h_300', '200x300', self::PNG_TEST_PORTRAIT_IMAGE],
+
             'Resize to width and height (landscape) square' =>
             ['w_300,h_150', '150x150', self::PNG_TEST_SQUARE_IMAGE],
             'Resize to width and height (landscape) landscape' =>
             ['w_300,h_150', '225x150', self::PNG_TEST_LANDSCAPE_IMAGE],
             'Resize to width and height (landscape) portrait' =>
             ['w_300,h_150', '100x150', self::PNG_TEST_PORTRAIT_IMAGE],
+
             'Resize to width and height (portrait) square' =>
             ['w_150,h_300', '150x150', self::PNG_TEST_SQUARE_IMAGE],
             'Resize to width and height (portrait) landscape' =>
             ['w_150,h_300', '150x100', self::PNG_TEST_LANDSCAPE_IMAGE],
             'Resize to width and height (portrait) portrait' =>
             ['w_150,h_300', '150x225', self::PNG_TEST_PORTRAIT_IMAGE],
+
             'Resize and Crop to square a square' =>
             ['w_300,h_300,c_1', '300x300', self::PNG_TEST_SQUARE_IMAGE],
             'Resize and Crop to square a landscape' =>
             ['w_300,h_300,c_1', '300x300', self::PNG_TEST_LANDSCAPE_IMAGE],
             'Resize and Crop to square a portrait' =>
             ['w_300,h_300,c_1', '300x300', self::PNG_TEST_PORTRAIT_IMAGE],
+
             'Resize and Crop to portrait (wider than portrait) square' =>
             ['w_250,h_300,c_1', '250x300', self::PNG_TEST_SQUARE_IMAGE],
             'Resize and Crop to portrait (wider than portrait) landscape' =>
             ['w_250,h_300,c_1', '250x300', self::PNG_TEST_LANDSCAPE_IMAGE],
             'Resize and Crop to portrait (wider than portrait) portrait' =>
             ['w_250,h_300,c_1', '250x300', self::PNG_TEST_PORTRAIT_IMAGE],
+
             'Resize and Crop to portrait (narrower than portrait) square' =>
             ['w_150,h_300,c_1', '150x300', self::PNG_TEST_SQUARE_IMAGE],
             'Resize and Crop to portrait (narrower than portrait) landscape' =>
             ['w_150,h_300,c_1', '150x300', self::PNG_TEST_LANDSCAPE_IMAGE],
             'Resize and Crop to portrait (narrower than portrait) portrait' =>
             ['w_150,h_300,c_1', '150x300', self::PNG_TEST_PORTRAIT_IMAGE],
+
             'Resize and Crop to landscape (taller than landscape) square' =>
             ['w_300,h_250,c_1', '300x250', self::PNG_TEST_SQUARE_IMAGE],
             'Resize and Crop to landscape (taller than landscape) landscape' =>
             ['w_300,h_250,c_1', '300x250', self::PNG_TEST_LANDSCAPE_IMAGE],
             'Resize and Crop to landscape (taller than landscape) portrait' =>
             ['w_300,h_250,c_1', '300x250', self::PNG_TEST_PORTRAIT_IMAGE],
+
             'Resize and Crop to landscape (shorter than landscape) square' =>
             ['w_300,h_150,c_1', '300x150', self::PNG_TEST_SQUARE_IMAGE],
             'Resize and Crop to landscape (shorter than landscape) landscape' =>
@@ -191,6 +199,14 @@ class ImageProcessorTest extends BaseTest
             'Expand and Crop to landscape (shorter a.r. than landscape) portrait' =>
             ['w_600,h_300,c_1', '200x300', self::PNG_TEST_SMALL_PORTRAIT_IMAGE],
 
+        ];
+
+        return $this->addOutputExtensionsToTests(array_merge($resizingTests, $this->partialCropTestProvider()));
+    }
+
+    protected function partialCropTestProvider(): array
+    {
+        return [
             // test partial crops without expanding
             'Expand and partial crop to square a landscape' =>
             ['w_250,h_250,c_1', '250x200', self::PNG_TEST_SMALL_LANDSCAPE_IMAGE],
@@ -224,10 +240,7 @@ class ImageProcessorTest extends BaseTest
             ['w_320,h_190,c_1', '300x190', self::PNG_TEST_SMALL_LANDSCAPE_IMAGE],
             'Expand and partial crop to landscape (shorter a.r. than landscape) portrait' =>
             ['w_320,h_190,c_1', '200x190', self::PNG_TEST_SMALL_PORTRAIT_IMAGE],
-
         ];
-
-        return $this->addOutputExtensionsToTests($resizingTests);
     }
 
     protected function addOutputExtensionsToTests(array $transformationsList): array
